move testcase
