+++ src/test/java/spoon/test/ctType/CtTypeTest.java
@@ -37,7 +37,6 @@ import spoon.reflect.visitor.filter.NamedElementFilter;
 import spoon.reflect.visitor.filter.TypeFilter;
 import spoon.test.ctType.testclasses.X;
 
-import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -46,6 +45,9 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
+import static org.hamcrest.CoreMatchers.anyOf;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertSame;
@@ -266,5 +268,20 @@ public class CtTypeTest {
 		CtType<?> reFetchedTypeDecl = typeRef.getTypeDeclaration();
 
 		assertSame(reFetchedTypeDecl, typeDecl);
+	}
+	@Test
+	public void testGetAllExecutablesOnTypeImplementingNestedInterface() {
+		// contract: implicit static nested interfaces are correct handled in getAllExecutables.
+		Launcher launcher = new Launcher();
+		launcher.addInputResource("src/test/resources/extendsStaticInnerType");
+		CtModel model = launcher.buildModel();
+		CtType<?> type = model.getAllTypes().stream().filter(v -> v.getSimpleName().contains("BarBaz")).findAny().get();
+		int expectedNumExecutablesInJDK8 = 13;
+		int expectedNumExecutablesPostJDK8 = 14;
+		int numExecutables = type.getAllExecutables().size();
+		assertThat(numExecutables, anyOf(
+				equalTo(expectedNumExecutablesInJDK8),
+				equalTo(expectedNumExecutablesPostJDK8))
+		);	
 	}
 }
